Updates dashboard and docs (#795)
* New site

* Updates site previews

* Fixes previews

* Updates mobile layout

* Add favicon generation

* Improves mobile display

* Updates todo

* Fixes domain setup flow

* Fixes hardcoding of host

* Update site previews

* Updates template editor

* Disable spec

* Updates template editor

* Cleans up

* Smoothes sign up flow

* Updates template editor

* Streamlines dashboard build

* Consolidate CDN URL helper

* Major consolidation

* Cleanup

* Updates dash

* Update dashboard

* Updates package

* Clean up

* Updates from email

* Removes FAST flag

* Fix

* Fix

* Adds integration tests

* Removes integration tests

* Revert changes to Dockerfile

* Update site

* More tests

* Updates tests

* Updates broken link tests

* Updates broken link tests

* Fixes broken links

* Updates broken link checker

* Adds broken link check for dashboard

* Fixes broken links in dashboard

* Tests can now run locally alongside server

* Removes AI bullshit

* Updates tests

* Updates log in tests

* Adds pupeteer tests

* Adds sign up test

* Updates sign up test

* Updates sign up test

* Refactor setup tests

* Update CI

* Updates workflows

* Updates sign up route

* Updates workflows

* Updates ci

* Updates tests

* Updates tests

* Updates settings

* Updates integration tests

* move login into setup

* Fixes test suite

* Updates test dockerfile

* Adds script to detect unused CSS

* Updates CSS generation

* Move around CSS

* More CSS cleaning

* Cleans up CSS

* Refactor CSS

* Updates CSS

* Updates unused CSS

* Updates tests:

* Updates tests

* Updates tests

* Adds more logging

* Updates featured sites loader

* Updates Dockerfile

* Updates commits logging

* Updates git workflow

* Further tweak

* Updates tools

* Cleans up CSS

* More clean up

* Updates docs

* Updates dashboard

+++ app/documentation/build/index.js
@@ -1,30 +1,50 @@
 const config = require("config");
-const { join } = require("path");
-const { build } = require("esbuild");
+const { join, dirname, basename, extname } = require("path");
 const fs = require("fs-extra");
 const chokidar = require("chokidar");
 const html = require("./html");
-const SOURCE_DIRECTORY = join(config.blot_directory, "/app/views");
-const DESTINATION_DIRECTORY = join(
-  config.blot_directory,
-  "/app/documentation/data"
-);
+const favicon = require("./favicon");
+const recursiveReadDir = require("../../helper/recursiveReadDirSync");
+
+const SOURCE_DIRECTORY = join(__dirname, "../../views");
+const DESTINATION_DIRECTORY = join(__dirname, "../data");
+
+const buildCSS = require("./css")({source: SOURCE_DIRECTORY, destination: DESTINATION_DIRECTORY});
+const buildJS = require("./js")({source: SOURCE_DIRECTORY, destination: DESTINATION_DIRECTORY});
+
 const zip = require("templates/folders/zip");
 const tools = require("./tools");
+const generateThumbnail = require("./generate-thumbnail");
 
-async function handle (path) {
+const handle = (initial = false) => async (path) => {
   try {
-    if (path.includes("images/examples")) {
-      await generateThumbnail(path);
+    
+    if (path.includes("tools/") && !initial) {
+      await tools();
+    } 
+    
+    if (path.includes("images/examples") && path.endsWith(".png")) {
+      await fs.copy(
+        join(SOURCE_DIRECTORY, path),
+        join(DESTINATION_DIRECTORY, path)
+      );
+      await generateThumbnail(
+        join(SOURCE_DIRECTORY, path),
+        join(
+          DESTINATION_DIRECTORY,
+          dirname(path),
+          basename(path, extname(path)) + "-thumb.png"
+        )
+      );
     } else if (path.endsWith(".html") && !path.includes("dashboard/")) {
       await buildHTML(path);
-    } else if (path.endsWith(".css")) {
+    } else if (path.endsWith(".css") && !initial) {
       await fs.copy(
         join(SOURCE_DIRECTORY, path),
         join(DESTINATION_DIRECTORY, path)
       );
       await buildCSS();
-    } else if (path.endsWith(".js")) {
+    } else if (path.endsWith(".js") && !initial) {
       await fs.copy(
         join(SOURCE_DIRECTORY, path),
         join(DESTINATION_DIRECTORY, path)
@@ -36,40 +56,49 @@ async function handle (path) {
         join(DESTINATION_DIRECTORY, path)
       );
     }
-  } catch (e) {}
+  } catch (e) {
+    console.error(e);
+  }
 }
 
 module.exports = async ({ watch = false } = {}) => {
-  await fs.emptyDir(DESTINATION_DIRECTORY);
+  console.time("build");
+
+  // we only reset the destination directory in production
+  if (config.environment !== "development") {
+    await fs.emptyDir(DESTINATION_DIRECTORY);
+  } 
 
   await zip();
 
-  // recursively read every file in the source directory
-  const list = dir => {
-    const files = fs.readdirSync(dir);
-    return files.reduce((acc, file) => {
-      const path = join(dir, file);
-      const isDirectory = fs.statSync(path).isDirectory();
-      return isDirectory ? [...acc, ...list(path)] : [...acc, path];
-    }, []);
-  };
+  await favicon(join(SOURCE_DIRECTORY, "images/logo.svg"), join(DESTINATION_DIRECTORY, "favicon.ico"));
 
-  const paths = list(SOURCE_DIRECTORY).map(path =>
+  const paths = recursiveReadDir(SOURCE_DIRECTORY).map(path =>
     path.slice(SOURCE_DIRECTORY.length + 1)
   );
 
-  await Promise.all(paths.map(handle));
+  const initialHandler = handle(true);
+
+  await Promise.all(paths.map(initialHandler));
 
   await tools();
 
+  await buildCSS();
+
+  await buildJS();
+
+  console.timeEnd("build");
+
   if (watch) {
+    const handler = handle();
+
     chokidar
       .watch(SOURCE_DIRECTORY, {
         cwd: SOURCE_DIRECTORY,
         ignoreInitial: true
       })
       .on("all", async (event, path) => {
-        if (path) handle(path);
+        if (path) handler(path);
       });
   }
 };
@@ -79,56 +108,6 @@ async function buildHTML (path) {
   const result = await html(contents);
 
   await fs.outputFile(join(DESTINATION_DIRECTORY, path), result);
-}
-
-async function buildCSS () {
-  // merge all css files together into one file
-  const cssDir = join(SOURCE_DIRECTORY, "css");
-  const cssFiles = (await fs.readdir(cssDir)).filter(i => i.endsWith(".css"));
-  const cssContents = await Promise.all(
-    cssFiles.map(name => fs.readFile(join(cssDir, name), "utf-8"))
-  );
-
-  const mergedCSS = cssContents.join("\n\n");
-
-  await fs.writeFile(join(DESTINATION_DIRECTORY, "css.min.css"), mergedCSS);
-}
-
-async function buildJS () {
-  await build({
-    entryPoints: [join(SOURCE_DIRECTORY, "js/documentation.js")],
-    bundle: true,
-    minify: true,
-    // sourcemap: true,
-    target: ["chrome58", "firefox57", "safari11", "edge16"],
-    outfile: join(DESTINATION_DIRECTORY, "documentation.min.js")
-  });
-}
-
-const sharp = require("sharp");
-const { dirname, basename, extname } = require("path");
-
-async function generateThumbnail (path) {
-  if (path.includes("-")) return;
-
-  // copy the file as-is too
-  await fs.copy(
-    join(SOURCE_DIRECTORY, path),
-    join(DESTINATION_DIRECTORY, path)
-  );
-
-  if (path.includes(".mobile.")) return;
-
-  // generate a thumbnail
-  await sharp(join(SOURCE_DIRECTORY, path))
-    .resize({ width: 130 })
-    .toFile(
-      join(
-        DESTINATION_DIRECTORY,
-        dirname(path),
-        basename(path, extname(path)) + "-thumb.png"
-      )
-    );
 }
 
 if (require.main === module) {
